Refactor, squash bugs, improve docs.
This is a major refactor. The kubelet is now the only
Kubernetes component that is downloaded and started as
a binary. The kubelet is then used to bootstrap all
other kubernetes components.  Bugs have been fixed in
the kube-dns manifest.  Docs have been significantly
updated to reflect things that changed in the refactor
and to also include a list of known issues and a
roadmap.